pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamshape"
version = "0.1.0"
description = "A unified LLM response parser with streaming support for multiple providers"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Saikethan"}
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.7.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
    "litellm>=1.45.0",
]

[project.urls]
Homepage = "https://github.com/saikethan27/StreamShape"

[project.optional-dependencies]
openai = ["openai>=1.52.0"]
anthropic = ["anthropic>=0.34.0"]
google = ["google-genai>=0.6.0"]
test = [
    "pytest>=8.3.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0.0",
]
all = [
    "openai>=1.52.0",
    "anthropic>=0.34.0",
    "google-genai>=0.6.0",
    "litellm>=1.45.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
